Changed the formula for distance

+++ Distance_Between_Two_Cities/Location.cs
@@ -9,9 +9,9 @@ namespace Distance_Between_Two_Cities
     class Location
     {
         //latitude and longitude
-        public float xCord { get; set; }
-        public float yCord { get; set; }
-        public Location(float x ,float y)
+        public double xCord { get; set; }
+        public double yCord { get; set; }
+        public Location(double x ,double y)
         {
             this.xCord = x;
             this.yCord = y;
